requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
claude-agent-sdk>=0.1.0
websockets>=12.0
